shipments/requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.20
django-celery-beat==2.8.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
greenlet==3.2.1
idna==3.10
kombu==5.5.3
playwright==1.51.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pyee==12.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
